+++ src/utils.ts
@@ -235,9 +235,6 @@ export function arrify<T>(val: T | T[]) {
 }
 
 export function ensureProgram(instance: TSInstance) {
-  if (instance.solutionBuilder) {
-    instance.solutionBuilder.buildReferences(instance.configFilePath!);
-  }
   if (instance && instance.watchHost) {
     if (instance.hasUnaccountedModifiedFiles) {
       if (instance.changedFilesList) {
@@ -358,11 +355,14 @@ export function validateSourceMapOncePerProject(
   }
 }
 
-export function supportsSolutionBuild(
-  loaderOptions: LoaderOptions,
-  compiler: typeof typescript
-) {
-  return !!loaderOptions.projectReferences && !!compiler.InvalidatedProjectKind;
+export function supportsSolutionBuild(instance: TSInstance) {
+  return (
+    !!instance.configFilePath &&
+    !!instance.loaderOptions.projectReferences &&
+    !!instance.compiler.InvalidatedProjectKind &&
+    !!instance.configParseResult.projectReferences &&
+    !!instance.configParseResult.projectReferences.length
+  );
 }
 
 /**