+++ test/comparison-tests/projectReferencesMultiple/patch1/indirectWithError/fileWithError.ts
@@ -1,0 +1,3 @@
+export function foo(): string {
+    return "hello i fixed this error";
+}