+++ test/comparison-tests/resolveJsonModule/webpack.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  entry: './app.ts',
+  output: {
+    filename: 'bundle.js'
+  },
+  resolve: {
+    extensions: ['.ts', '.json']
+  },
+  module: {
+    rules: [{ test: /\.tsx?$/, loader: 'ts-loader' }]
+  }
+};